refactor: move MegaHeader link helper
